+++ src/settings.ts
@@ -7,8 +7,10 @@ export const PAIRING_SERVICE_PORT = 6467;
 export const STORAGE_DIRECTORY = 'androidtv-ultimate';
 export const CREDENTIALS_FILE = 'credentials.json';
 export const STATUS_FILE = 'status.json';
+export const DISCOVERY_FILE = 'discovery.json';
 
 export const CLIENT_MODEL = 'Homebridge';
 export const CLIENT_VENDOR = 'Tharun P Karun';
 export const CLIENT_PACKAGE = 'homebridge-androidtv-ultimate';
 export const PROTOCOL_VERSION = 622;
+export const DEFAULT_DISCOVERY_INTERVAL_SECONDS = 60;